feat(BaseModel): add annotations to BaseModel
With this change all the objects now can have annotations.
